[WiP] Bug 7826: Data validation failed for path
Modified removeGroup entity and removeFlow entity method
to fix the CSIT error of group-id not existing.

Will run the CSIT and check.

+++ mdsalutil/mdsalutil-impl/src/main/java/org/opendaylight/genius/mdsalutil/internal/MDSALManager.java
@@ -179,41 +179,28 @@ public class MDSALManager extends AbstractLifecycle implements IMdsalApiManager 
         }
 
         FlowKey flowKey = new FlowKey(new FlowId(flowEntity.getFlowId()));
-
         FlowBuilder flowbld = flowEntity.getFlowBuilder();
-
-        Node nodeDpn = buildDpnNode(flowEntity.getDpnId());
-        InstanceIdentifier<Flow> flowInstanceId = InstanceIdentifier.builder(Nodes.class)
-                .child(Node.class, nodeDpn.getKey()).augmentation(FlowCapableNode.class)
-                .child(Table.class, new TableKey(flowEntity.getTableId())).child(Flow.class, flowKey).build();
-
+        short tableId = flowEntity.getTableId();
+        InstanceIdentifier<Flow> flowInstanceId = buildFlowInstanceIdentifier(flowEntity.getDpnId(), tableId,
+                    flowKey);
         tx.put(LogicalDatastoreType.CONFIGURATION, flowInstanceId, flowbld.build(), true);
     }
 
     public void writeFlowInternal(BigInteger dpId, Flow flow, WriteTransaction tx) {
         FlowKey flowKey = new FlowKey(new FlowId(flow.getId()));
-        Node nodeDpn = buildDpnNode(dpId);
-        InstanceIdentifier<Flow> flowInstanceId = InstanceIdentifier.builder(Nodes.class)
-                .child(Node.class, nodeDpn.getKey()).augmentation(FlowCapableNode.class)
-                .child(Table.class, new TableKey(flow.getTableId())).child(Flow.class, flowKey).build();
+        InstanceIdentifier<Flow> flowInstanceId = buildFlowInstanceIdentifier(dpId, flow.getTableId(), flowKey);
         tx.put(LogicalDatastoreType.CONFIGURATION, flowInstanceId, flow, true);
     }
 
     public void batchedAddFlowInternal(BigInteger dpId, Flow flow) {
         FlowKey flowKey = new FlowKey(new FlowId(flow.getId()));
-        Node nodeDpn = buildDpnNode(dpId);
-        InstanceIdentifier<Flow> flowInstanceId = InstanceIdentifier.builder(Nodes.class)
-                .child(Node.class, nodeDpn.getKey()).augmentation(FlowCapableNode.class)
-                .child(Table.class, new TableKey(flow.getTableId())).child(Flow.class, flowKey).build();
+        InstanceIdentifier<Flow> flowInstanceId = buildFlowInstanceIdentifier(dpId, flow.getTableId(), flowKey);
         FlowBatchingUtils.write(flowInstanceId, flow);
     }
 
     public void batchedRemoveFlowInternal(BigInteger dpId, Flow flow) {
         FlowKey flowKey = new FlowKey(new FlowId(flow.getId()));
-        Node nodeDpn = buildDpnNode(dpId);
-        InstanceIdentifier<Flow> flowInstanceId = InstanceIdentifier.builder(Nodes.class)
-                .child(Node.class, nodeDpn.getKey()).augmentation(FlowCapableNode.class)
-                .child(Table.class, new TableKey(flow.getTableId())).child(Flow.class, flowKey).build();
+        InstanceIdentifier<Flow> flowInstanceId = buildFlowInstanceIdentifier(dpId, flow.getTableId(), flowKey);
         FlowBatchingUtils.delete(flowInstanceId);
     }
 
@@ -272,8 +259,11 @@ public class MDSALManager extends AbstractLifecycle implements IMdsalApiManager 
         Node nodeDpn = buildDpnNode(dpId);
         long groupId = group.getGroupId().getValue();
         InstanceIdentifier<Group> groupInstanceId = buildGroupInstanceIdentifier(groupId, nodeDpn);
-
-        tx.delete(LogicalDatastoreType.CONFIGURATION, groupInstanceId);
+        if (groupExists(dpId, groupId)) {
+            tx.delete(LogicalDatastoreType.CONFIGURATION, groupInstanceId);
+        } else {
+            LOG.warn("Group {} does not exist", groupId);
+        }
     }
 
     public CheckedFuture<Void, TransactionCommitFailedException> removeFlowInternal(FlowEntity flowEntity) {
@@ -308,13 +298,14 @@ public class MDSALManager extends AbstractLifecycle implements IMdsalApiManager 
     }
 
     public void deleteFlowEntityInternal(FlowEntity flowEntity, WriteTransaction tx) {
-        Node nodeDpn = buildDpnNode(flowEntity.getDpnId());
+        short tableId = flowEntity.getTableId();
         FlowKey flowKey = new FlowKey(new FlowId(flowEntity.getFlowId()));
-        InstanceIdentifier<Flow> flowInstanceId = InstanceIdentifier.builder(Nodes.class)
-                .child(Node.class, nodeDpn.getKey()).augmentation(FlowCapableNode.class)
-                .child(Table.class, new TableKey(flowEntity.getTableId())).child(Flow.class, flowKey).build();
-
-        tx.delete(LogicalDatastoreType.CONFIGURATION, flowInstanceId);
+        InstanceIdentifier<Flow> flowInstanceId = buildFlowInstanceIdentifier(flowEntity.getDpnId(), tableId, flowKey);
+        if (flowExists(flowEntity.getDpnId(), tableId, flowKey)) {
+            tx.delete(LogicalDatastoreType.CONFIGURATION, flowInstanceId);
+        } else {
+            LOG.warn("Flow {} does not exist", flowKey);
+        }
     }
 
     public CheckedFuture<Void, TransactionCommitFailedException> removeFlowNewInternal(BigInteger dpnId,
@@ -328,10 +319,13 @@ public class MDSALManager extends AbstractLifecycle implements IMdsalApiManager 
     public void deleteFlowInternal(BigInteger dpId, Flow flow, WriteTransaction tx) {
         Node nodeDpn = buildDpnNode(dpId);
         FlowKey flowKey = new FlowKey(flow.getId());
-        InstanceIdentifier<Flow> flowInstanceId = InstanceIdentifier.builder(Nodes.class)
-                .child(Node.class, nodeDpn.getKey()).augmentation(FlowCapableNode.class)
-                .child(Table.class, new TableKey(flow.getTableId())).child(Flow.class, flowKey).build();
-        tx.delete(LogicalDatastoreType.CONFIGURATION, flowInstanceId);
+        short tableId = flow.getTableId();
+        InstanceIdentifier<Flow> flowInstanceId = buildFlowInstanceIdentifier(dpId, tableId, flowKey);
+        if (flowExists(dpId, tableId, flowKey)) {
+            tx.delete(LogicalDatastoreType.CONFIGURATION, flowInstanceId);
+        } else {
+            LOG.warn("Flow {} does not exist", flowKey);
+        }
     }
 
     public CheckedFuture<Void, TransactionCommitFailedException> removeGroupInternal(GroupEntity groupEntity) {
@@ -366,9 +360,13 @@ public class MDSALManager extends AbstractLifecycle implements IMdsalApiManager 
 
     public void removeGroupEntityInternal(GroupEntity groupEntity, WriteTransaction tx) {
         Node nodeDpn = buildDpnNode(groupEntity.getDpnId());
-        InstanceIdentifier<Group> groupInstanceId = buildGroupInstanceIdentifier(groupEntity.getGroupId(), nodeDpn);
-
-        tx.delete(LogicalDatastoreType.CONFIGURATION, groupInstanceId);
+        long groupId = groupEntity.getGroupId();
+        InstanceIdentifier<Group> groupInstanceId = buildGroupInstanceIdentifier(groupId, nodeDpn);
+        if (groupExists(nodeDpn, groupId)) {
+            tx.delete(LogicalDatastoreType.CONFIGURATION, groupInstanceId);
+        } else {
+            LOG.warn("Group {} does not exist", groupId);
+        }
     }
 
     public void modifyGroupInternal(GroupEntity groupEntity) {
@@ -434,13 +432,14 @@ public class MDSALManager extends AbstractLifecycle implements IMdsalApiManager 
         }
         Flow flow = flowEntity.getFlowBuilder().build();
         String flowId = flowEntity.getFlowId();
+        short tableId = flowEntity.getTableId();
         BigInteger dpId = flowEntity.getDpnId();
         FlowKey flowKey = new FlowKey(new FlowId(flowId));
         Node nodeDpn = buildDpnNode(dpId);
-        InstanceIdentifier<Flow> flowInstanceId = InstanceIdentifier.builder(Nodes.class)
-                .child(Node.class, nodeDpn.getKey()).augmentation(FlowCapableNode.class)
-                .child(Table.class, new TableKey(flow.getTableId())).child(Flow.class, flowKey).build();
-        if (isRemove) {
+        InstanceIdentifier<Flow> flowInstanceId = buildFlowInstanceIdentifier(dpId, flow.getTableId(), flowKey);
+        Optional<Flow> flowExists = MDSALUtil.read(dataBroker, LogicalDatastoreType.CONFIGURATION,
+                flowInstanceId);
+        if (isRemove && flowExists(dpId, tableId, flowKey)) {
             MDSALUtil.syncDelete(dataBroker, LogicalDatastoreType.CONFIGURATION, flowInstanceId);
         } else {
             MDSALUtil.syncWrite(dataBroker, LogicalDatastoreType.CONFIGURATION, flowInstanceId, flow);
@@ -456,7 +455,9 @@ public class MDSALManager extends AbstractLifecycle implements IMdsalApiManager 
         Node nodeDpn = buildDpnNode(dpId);
         long groupId = groupEntity.getGroupId();
         InstanceIdentifier<Group> groupInstanceId = buildGroupInstanceIdentifier(groupEntity.getGroupId(), nodeDpn);
-        if (isRemove) {
+        Optional<Group> groupExists = MDSALUtil.read(dataBroker, LogicalDatastoreType.CONFIGURATION,
+                groupInstanceId);
+        if (isRemove && groupExists(dpId, groupId)) {
             MDSALUtil.syncDelete(dataBroker, LogicalDatastoreType.CONFIGURATION, groupInstanceId);
         } else {
             MDSALUtil.syncWrite(dataBroker, LogicalDatastoreType.CONFIGURATION, groupInstanceId, group);
@@ -468,7 +469,9 @@ public class MDSALManager extends AbstractLifecycle implements IMdsalApiManager 
         Node nodeDpn = buildDpnNode(dpId);
         long groupId = group.getGroupId().getValue();
         InstanceIdentifier<Group> groupInstanceId = buildGroupInstanceIdentifier(groupId, nodeDpn);
-        if (isRemove) {
+        Optional<Group> groupExists = MDSALUtil.read(dataBroker, LogicalDatastoreType.CONFIGURATION,
+                groupInstanceId);
+        if (isRemove && groupExists(dpId, groupId)) {
             MDSALUtil.syncDelete(dataBroker, LogicalDatastoreType.CONFIGURATION, groupInstanceId);
         } else {
             MDSALUtil.syncWrite(dataBroker, LogicalDatastoreType.CONFIGURATION, groupInstanceId, group);
@@ -724,6 +727,8 @@ public class MDSALManager extends AbstractLifecycle implements IMdsalApiManager 
 
         if (groupExists(nodeDpn, groupId)) {
             tx.delete(LogicalDatastoreType.CONFIGURATION, bucketInstanceId);
+        } else {
+            LOG.warn("Group {} does not exist", groupId);
         }
     }
 
@@ -749,7 +754,7 @@ public class MDSALManager extends AbstractLifecycle implements IMdsalApiManager 
                     groupInstanceId);
             return groupOptional.isPresent();
         } catch (ReadFailedException e) {
-            LOG.warn("Exception while reading group {} for Node {}", nodeDpn.getKey(), groupId);
+            LOG.warn("Exception while reading group {} for Node {}", groupId, nodeDpn.getKey());
         }
         return false;
     }
@@ -759,6 +764,26 @@ public class MDSALManager extends AbstractLifecycle implements IMdsalApiManager 
                 .child(Node.class, nodeDpn.getKey()).augmentation(FlowCapableNode.class)
                 .child(Group.class, new GroupKey(new GroupId(groupId))).build();
         return groupInstanceId;
+    }
+
+    public boolean flowExists(BigInteger dpId, short tableId, FlowKey flowKey) {
+        InstanceIdentifier<Flow> flowInstanceId = buildFlowInstanceIdentifier(dpId, tableId, flowKey);
+        try {
+            Optional<Flow> flowOptional = singleTxDb.syncReadOptional(LogicalDatastoreType.CONFIGURATION,
+                    flowInstanceId);
+            return flowOptional.isPresent();
+        } catch (ReadFailedException e) {
+            LOG.warn("Exception while reading flow {} for dpn {}", flowKey, dpId);
+        }
+        return false;
+    }
+
+    private InstanceIdentifier<Flow> buildFlowInstanceIdentifier(BigInteger dpnId, short tableId, FlowKey flowKey) {
+        Node nodeDpn = buildDpnNode(dpnId);
+        InstanceIdentifier<Flow> flowInstanceId = InstanceIdentifier.builder(Nodes.class)
+                .child(Node.class, nodeDpn.getKey()).augmentation(FlowCapableNode.class)
+                .child(Table.class, new TableKey(tableId)).child(Flow.class, flowKey).build();
+        return flowInstanceId;
     }
 
     private InstanceIdentifier<Bucket> buildBucketInstanceIdentifier(long groupId, long bucketId,
